KO-224: integer stale threshold, unknown hosts stay unknown, bold-optional What/Why/How
- sweep_config() returns heartbeat_stale_ms as an int, so /status
  serialises 300000 rather than 300000.0.
- host_label() returns host_name(host) for a None host even under a
  [report] host_label: a pre-column run has no recorded host, and the
  report shows ? while /status and /runs show null.
- BOLD_KEY_RE accepts What:, **What:** and **What: ** (and Why/How) as
  one key; the bold run is a Linear serialisation detail, not a contract.

Corrects the two assertions that pinned the old behaviour and adds
witnesses for each surface.

+++ holophyte/config.py
@@ -440,7 +440,7 @@ def sweep_config(target):
                 f"{constraint}, got {value!r}")
         values[key] = value
     return SweepConfig(
-        heartbeat_stale_ms=values["heartbeat_stale_min"] * 60000,
+        heartbeat_stale_ms=int(values["heartbeat_stale_min"] * 60000),
         stale_strikes=values["stale_strikes"],
         budget_grace=values["budget_grace"],
         review_overlap_threshold=values["review_overlap_threshold"],
